add multiple jwtconfig
