+++ dpppt-backend-sdk/dpppt-backend-sdk-ws/src/main/resources/generated_pub.pem
@@ -1,0 +1,1 @@
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAiy5bG00o0iBAoiXE3dVUbFl2sppZz7Bzo+zukYBHNHVnQHcmv4BwEN2Uwch5SeWHHf8mDJnv1Jo2cvOaJG0Sp0/mRl4WVNK1mCIYfGUpiAeWF+kypVZxhk0su7Oj5p8u3wRZvm859BonUB2strLcd3N08AfTVko0n0qbyfN6KyQdIq6RwVwR5JMYiHOn9IHl+s+wXczjqUTocdbGtWYhmT2dXyOc3wHEa1tln6kQnFJNczXuWR3fZWtR/GLPQlHS3+15mV1JunjD82IHJxIuKH0OngAtny/KkOPMqxehILJUS27Tuhs9gW6xSY0leRJ8oNHBAxrCW0pJlyI4j5mUoQIDAQAB
